refactor to go style project
