+++ internal/scan/podman.go
@@ -1,4 +1,4 @@
-package main
+package scan
 
 import (
 	"bytes"
