+++ internal/scan/types_config.go
@@ -1,4 +1,4 @@
-package main
+package scan
 
 import (
 	"strings"
