+++ internal/scan/types_scan_result.go
@@ -1,4 +1,4 @@
-package main
+package scan
 
 import (
 	"errors"
