+++ internal/scan/types_scan_results.go
@@ -1,4 +1,4 @@
-package main
+package scan
 
 func NewScanResults() *ScanResults {
 	return &ScanResults{}
