+++ internal/scan/validation_test.go
@@ -1,4 +1,4 @@
-package main
+package scan
 
 import (
 	"context"
