+++ internal/scan/validations_openssl.go
@@ -1,4 +1,4 @@
-package main
+package scan
 
 import (
 	"bytes"
